+++ test/e2e-runner/src/linux.rs
@@ -47,5 +47,8 @@ pub fn run_linux_e2e() {
         LinuxScenarios::FleetControlApi(args) => {
             fleet_control_api::run_fleet_control_api(&args.fleet_control);
         }
+        LinuxScenarios::SelfUpdateFromLatest(args) => {
+            scenarios::self_update::test_self_update_from_latest(args);
+        }
     };
 }